MovingSprite 상속하는 Player 클래스 만들기.

+++ cpp_18/Cpp18/Ball.cpp
@@ -4,6 +4,6 @@
 Ball::Ball(Texture& texture)
 {
 	setTexture(texture);
-	vector.x = Game::getRandom(100) + 50;
-	vector.y = Game::getRandom(100) + 20;
+	vector.x = (float)(Game::getRandom(100) + 50);
+	vector.y = (float)(Game::getRandom(100) + 20);
 }
